+++ examples/szengine/a_header_szengine.py
@@ -7,18 +7,14 @@ Simply a header used in development.
 from typing import List, Tuple
 
 import grpc
+from senzing import SZ_WITHOUT_INFO, SzEngine
 from senzing_truthset import (
     TRUTHSET_CUSTOMER_RECORDS,
     TRUTHSET_REFERENCE_RECORDS,
     TRUTHSET_WATCHLIST_RECORDS,
 )
 
-from senzing_grpc import (
-    SZ_WITHOUT_INFO,
-    SzAbstractFactory,
-    SzAbstractFactoryParameters,
-    SzEngine,
-)
+from senzing_grpc import SzAbstractFactory, SzAbstractFactoryParameters
 
 DATA_SOURCES = {
     "CUSTOMERS": TRUTHSET_CUSTOMER_RECORDS,