+++ src/senzing_grpc/__init__.py
@@ -1,4 +1,4 @@
-from senzing_abstract import (
+from senzing import (
     ENGINE_EXCEPTION_MAP,
     SZ_INITIALIZE_WITH_DEFAULT_CONFIGURATION,
     SZ_NO_ATTRIBUTES,
@@ -10,25 +10,14 @@ from senzing_abstract import (
     SZ_VERBOSE_LOGGING,
     SZ_WITHOUT_INFO,
     SzBadInputError,
-)
-from senzing_abstract import SzConfigAbstract as SzConfig
-from senzing_abstract import SzConfigManagerAbstract as SzConfigManager
-from senzing_abstract import (
     SzConfigurationError,
     SzDatabaseConnectionLostError,
     SzDatabaseError,
-)
-from senzing_abstract import SzDiagnosticAbstract as SzDiagnostic
-from senzing_abstract import SzEngineAbstract as SzEngine
-from senzing_abstract import (
     SzEngineFlags,
     SzError,
     SzLicenseError,
     SzNotFoundError,
     SzNotInitializedError,
-)
-from senzing_abstract import SzProductAbstract as SzProduct
-from senzing_abstract import (
     SzReplaceConflictError,
     SzRetryableError,
     SzRetryTimeoutExceededError,
@@ -38,11 +27,11 @@ from senzing_abstract import (
 )
 
 from .szabstractfactory import SzAbstractFactory, SzAbstractFactoryParameters
-from .szconfig import SzConfig as SzConfigGrpc
-from .szconfigmanager import SzConfigManager as SzConfigManagerGrpc
-from .szdiagnostic import SzDiagnostic as SzDiagnosticGrpc
-from .szengine import SzEngine as SzEngineGrpc
-from .szproduct import SzProduct as SzProductGrpc
+from .szconfig import SzConfig
+from .szconfigmanager import SzConfigManager
+from .szdiagnostic import SzDiagnostic
+from .szengine import SzEngine
+from .szproduct import SzProduct
 
 __all__ = [
     "ENGINE_EXCEPTION_MAP",
@@ -50,7 +39,6 @@ __all__ = [
     "SZ_NO_ATTRIBUTES",
     "SZ_NO_AVOIDANCES",
     "SZ_NO_FLAGS",
-    "SZ_NO_LOGGING",
     "SZ_NO_LOGGING",
     "SZ_NO_REQUIRED_DATASOURCES",
     "SZ_NO_SEARCH_PROFILE",
@@ -60,23 +48,18 @@ __all__ = [
     "SzAbstractFactoryParameters",
     "SzBadInputError",
     "SzConfig",
-    "SzConfigGrpc",
     "SzConfigManager",
-    "SzConfigManagerGrpc",
     "SzConfigurationError",
     "SzDatabaseConnectionLostError",
     "SzDatabaseError",
     "SzDiagnostic",
-    "SzDiagnosticGrpc",
     "SzEngine",
-    "SzEngineGrpc",
     "SzEngineFlags",
     "SzError",
     "SzLicenseError",
     "SzNotFoundError",
     "SzNotInitializedError",
     "SzProduct",
-    "SzProductGrpc",
     "SzReplaceConflictError",
     "SzRetryableError",
     "SzRetryTimeoutExceededError",