+++ tests/szengine_test.py
@@ -13,12 +13,18 @@ from senzing_truthset import (
     TRUTHSET_WATCHLIST_RECORDS,
 )
 
-from senzing_grpc import SZ_NO_FLAGS, SZ_WITHOUT_INFO, SzBadInputError, SzConfig
-from senzing_grpc import SzConfigGrpc as SzConfigTest
-from senzing_grpc import SzConfigManagerGrpc as SzConfigManagerTest
-from senzing_grpc import SzEngineFlags
-from senzing_grpc import SzEngineGrpc as SzEngineTest
-from senzing_grpc import SzError, SzNotFoundError, SzUnknownDataSourceError
+from senzing_grpc import (
+    SZ_NO_FLAGS,
+    SZ_WITHOUT_INFO,
+    SzBadInputError,
+    SzConfig,
+    SzConfigManager,
+    SzEngine,
+    SzEngineFlags,
+    SzError,
+    SzNotFoundError,
+    SzUnknownDataSourceError,
+)
 
 DATA_SOURCES = {
     "CUSTOMERS": TRUTHSET_CUSTOMER_RECORDS,
@@ -63,9 +69,9 @@ RECORD_STR_BAD = (
 
 
 def test_add_truthset_datasources(
-    sz_engine: SzEngineTest,
-    sz_configmanager: SzConfigManagerTest,
-    sz_config: SzConfigTest,
+    sz_engine: SzEngine,
+    sz_configmanager: SzConfigManager,
+    sz_config: SzConfig,
 ) -> None:
     """Add needed datasources for tests."""
     config_handle = sz_config.create_config()
@@ -82,7 +88,7 @@ def test_add_truthset_datasources(
 # -----------------------------------------------------------------------------
 
 
-def test_add_record(sz_engine: SzEngineTest) -> None:
+def test_add_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().add_record()."""
     data_source_code = "TEST"
     record_id = "1"
@@ -91,7 +97,7 @@ def test_add_record(sz_engine: SzEngineTest) -> None:
     sz_engine.add_record(data_source_code, record_id, json.dumps(record_definition), flags)
 
 
-def test_add_record_bad_data_source_code_type(sz_engine: SzEngineTest) -> None:
+def test_add_record_bad_data_source_code_type(sz_engine: SzEngine) -> None:
     """Test SzEngine().add_record()."""
     bad_data_source_code = 1
     record_id = "1"
@@ -101,7 +107,7 @@ def test_add_record_bad_data_source_code_type(sz_engine: SzEngineTest) -> None:
         sz_engine.add_record(bad_data_source_code, record_id, record_definition, flags)  # type: ignore[arg-type]
 
 
-def test_add_record_bad_data_source_code_value(sz_engine: SzEngineTest) -> None:
+def test_add_record_bad_data_source_code_value(sz_engine: SzEngine) -> None:
     """Test SzEngine().add_record()."""
     bad_data_source_code = "DOESN'T EXIST"
     record_id = "1"
@@ -111,7 +117,7 @@ def test_add_record_bad_data_source_code_value(sz_engine: SzEngineTest) -> None:
         sz_engine.add_record(bad_data_source_code, record_id, json.dumps(record_definition), flags)
 
 
-def test_add_record_with_info(sz_engine: SzEngineTest) -> None:
+def test_add_record_with_info(sz_engine: SzEngine) -> None:
     """Test SzEngine().add_record_with_info()."""
     data_source_code = "TEST"
     record_id = "1"
@@ -122,7 +128,7 @@ def test_add_record_with_info(sz_engine: SzEngineTest) -> None:
     assert schema(add_record_with_info_schema) == actual_as_dict
 
 
-def test_add_record_bad_record(sz_engine: SzEngineTest) -> None:
+def test_add_record_bad_record(sz_engine: SzEngine) -> None:
     """Test add_record with bad JSON string."""
     data_source_code = "TEST"
     record_id = "1"
@@ -131,7 +137,7 @@ def test_add_record_bad_record(sz_engine: SzEngineTest) -> None:
         sz_engine.add_record(data_source_code, record_id, bad_record_definition)
 
 
-def test_add_record_bad_record_id_type(sz_engine: SzEngineTest) -> None:
+def test_add_record_bad_record_id_type(sz_engine: SzEngine) -> None:
     """Test add_record with incorrect record id type."""
     data_source_code = "TEST"
     bad_record_id = 1
@@ -140,7 +146,7 @@ def test_add_record_bad_record_id_type(sz_engine: SzEngineTest) -> None:
         sz_engine.add_record(data_source_code, bad_record_id, record_definition)  # type: ignore[arg-type]
 
 
-def test_add_record_data_source_code_empty(sz_engine: SzEngineTest) -> None:
+def test_add_record_data_source_code_empty(sz_engine: SzEngine) -> None:
     """Test add_record with empty data source code."""
     bad_data_source_code = ""
     record_id = "1"
@@ -149,7 +155,7 @@ def test_add_record_data_source_code_empty(sz_engine: SzEngineTest) -> None:
         sz_engine.add_record(bad_data_source_code, record_id, record_definition)
 
 
-def test_add_record_record_str_empty(sz_engine: SzEngineTest) -> None:
+def test_add_record_record_str_empty(sz_engine: SzEngine) -> None:
     """Test add_record with empty record as a string"""
     data_source_code = "TEST"
     record_id = "1"
@@ -165,7 +171,7 @@ def test_add_record_record_str_empty(sz_engine: SzEngineTest) -> None:
 #         sz_engine.add_record(data_source_code, record_id, {})
 
 
-def x_test_add_record_with_info_dict(sz_engine: SzEngineTest) -> None:
+def x_test_add_record_with_info_dict(sz_engine: SzEngine) -> None:
     """Test add_record with flag to return with_info where the record is a dict."""
     data_source_code = "TEST"
     record_id = "1"
@@ -176,7 +182,7 @@ def x_test_add_record_with_info_dict(sz_engine: SzEngineTest) -> None:
     assert schema(add_record_with_info_schema) == actual_as_dict
 
 
-def test_add_record_with_info_str(sz_engine: SzEngineTest) -> None:
+def test_add_record_with_info_str(sz_engine: SzEngine) -> None:
     """Test add_record with flag to return with_info where the record is a JSON string."""
     data_source_code = "TEST"
     record_id = "1"
@@ -189,7 +195,7 @@ def test_add_record_with_info_str(sz_engine: SzEngineTest) -> None:
 
 # TODO: Modify as_python_bytes to convert int to str? More robust and allows mistakes to continue
 def test_add_record_with_info_bad_data_source_code_type(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().add_record_with_info()."""
     bad_data_source_code = 1
@@ -201,7 +207,7 @@ def test_add_record_with_info_bad_data_source_code_type(
 
 
 def test_add_record_with_info_bad_data_source_code_value(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().add_record_with_info()."""
     bad_data_source_code = "DOESN'T EXIST"
@@ -212,7 +218,7 @@ def test_add_record_with_info_bad_data_source_code_value(
         _ = sz_engine.add_record(bad_data_source_code, record_id, json.dumps(record_definition), flags)
 
 
-def test_add_record_with_info_bad_record(sz_engine: SzEngineTest) -> None:
+def test_add_record_with_info_bad_record(sz_engine: SzEngine) -> None:
     """Test add_record with flag to return with_info with bad JSON string."""
     data_source_code = "TEST"
     record_id = "1"
@@ -222,7 +228,7 @@ def test_add_record_with_info_bad_record(sz_engine: SzEngineTest) -> None:
         sz_engine.add_record(data_source_code, record_id, bad_record_definition, flags)
 
 
-def test_add_record_with_info_bad_record_id_type(sz_engine: SzEngineTest) -> None:
+def test_add_record_with_info_bad_record_id_type(sz_engine: SzEngine) -> None:
     """Test add_record with flag to return with_info with incorrect record id type."""
     data_source_code = "TEST"
     bad_record_id = 1
@@ -232,7 +238,7 @@ def test_add_record_with_info_bad_record_id_type(sz_engine: SzEngineTest) -> Non
         sz_engine.add_record(data_source_code, bad_record_id, record_definition, flags)  # type: ignore[arg-type]
 
 
-def test_add_record_with_info_record_str_empty(sz_engine: SzEngineTest) -> None:
+def test_add_record_with_info_record_str_empty(sz_engine: SzEngine) -> None:
     """Test add_record_with_info with empty record as a string"""
     data_source_code = "TEST"
     record_id = "1"
@@ -246,13 +252,13 @@ def test_close_export() -> None:
     # TODO: implement.
 
 
-def test_count_redo_records(sz_engine: SzEngineTest) -> None:
+def test_count_redo_records(sz_engine: SzEngine) -> None:
     """Test SzEngine().count_redo_records()."""
     actual = sz_engine.count_redo_records()
     assert actual > 0
 
 
-def test_delete_record(sz_engine: SzEngineTest) -> None:
+def test_delete_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().delete_record()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -264,7 +270,7 @@ def test_delete_record(sz_engine: SzEngineTest) -> None:
     sz_engine.delete_record(data_source_code, record_id, flags)
 
 
-def test_delete_record_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_delete_record_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().delete_record()."""
     bad_data_source_code = "XXXX"
     record_id = "9999"
@@ -273,7 +279,7 @@ def test_delete_record_bad_data_source_code(sz_engine: SzEngineTest) -> None:
         sz_engine.delete_record(bad_data_source_code, record_id, flags)
 
 
-def test_delete_record_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_delete_record_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().delete_record()."""
     data_source_code = "CUSTOMERS"
     bad_record_id = "9999"
@@ -281,7 +287,7 @@ def test_delete_record_bad_record_id(sz_engine: SzEngineTest) -> None:
     sz_engine.delete_record(data_source_code, bad_record_id, flags)
 
 
-def test_delete_record_with_info(sz_engine: SzEngineTest) -> None:
+def test_delete_record_with_info(sz_engine: SzEngine) -> None:
     """Test SzEngine().delete_record_with_info()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -295,7 +301,7 @@ def test_delete_record_with_info(sz_engine: SzEngineTest) -> None:
     assert schema(add_record_with_info_schema) == actual_as_dict
 
 
-def test_delete_record_with_info_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_delete_record_with_info_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().delete_record_with_info()."""
     bad_data_source_code = "XXXX"
     record_id = "9999"
@@ -304,7 +310,7 @@ def test_delete_record_with_info_bad_data_source_code(sz_engine: SzEngineTest) -
         _ = sz_engine.delete_record(bad_data_source_code, record_id, flags)
 
 
-def test_delete_record_with_info_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_delete_record_with_info_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().delete_record_with_info()."""
     data_source_code = "CUSTOMERS"
     bad_record_id = "9999"
@@ -314,7 +320,7 @@ def test_delete_record_with_info_bad_record_id(sz_engine: SzEngineTest) -> None:
     assert schema(add_record_with_info_schema) == actual_as_dict
 
 
-def test_export_csv_entity_report(sz_engine: SzEngineTest) -> None:
+def test_export_csv_entity_report(sz_engine: SzEngine) -> None:
     """Test SzEngine().export_csv_entity_report()."""
     csv_column_list = "RESOLVED_ENTITY_ID,RESOLVED_ENTITY_NAME,RELATED_ENTITY_ID,MATCH_LEVEL,MATCH_KEY,IS_DISCLOSED,IS_AMBIGUOUS,DATA_SOURCE,RECORD_ID,JSON_DATA"
     flags = SzEngineFlags.SZ_EXPORT_DEFAULT_FLAGS
@@ -329,7 +335,7 @@ def test_export_csv_entity_report(sz_engine: SzEngineTest) -> None:
     assert len(actual) > 0
 
 
-def test_export_json_entity_report(sz_engine: SzEngineTest) -> None:
+def test_export_json_entity_report(sz_engine: SzEngine) -> None:
     """Test SzEngine().export_json_entity_report()."""
     handle = sz_engine.export_json_entity_report()
     actual = ""
@@ -350,7 +356,7 @@ def test_fetch_next() -> None:
     # TODO: implement test_fetch_next.
 
 
-def test_find_interesting_entities_by_entity_id(sz_engine: SzEngineTest) -> None:
+def test_find_interesting_entities_by_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_interesting_entities_by_entity_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -366,7 +372,7 @@ def test_find_interesting_entities_by_entity_id(sz_engine: SzEngineTest) -> None
 
 
 def test_find_interesting_entities_by_entity_id_bad_entity_id(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().find_interesting_entities_by_entity_id()."""
     bad_entity_id = 0
@@ -375,7 +381,7 @@ def test_find_interesting_entities_by_entity_id_bad_entity_id(
         _ = sz_engine.find_interesting_entities_by_entity_id(bad_entity_id, flags)
 
 
-def test_find_interesting_entities_by_record_id(sz_engine: SzEngineTest) -> None:
+def test_find_interesting_entities_by_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_interesting_entities_by_record_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -392,7 +398,7 @@ def test_find_interesting_entities_by_record_id(sz_engine: SzEngineTest) -> None
 
 
 def test_find_interesting_entities_by_record_id_bad_data_source_code(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().find_interesting_entities_by_record_id()."""
     bad_data_source_code = "XXXX"
@@ -403,7 +409,7 @@ def test_find_interesting_entities_by_record_id_bad_data_source_code(
 
 
 def test_find_interesting_entities_by_record_id_bad_record_id(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().find_interesting_entities_by_record_id()."""
     data_source_code = "CUSTOMERS"
@@ -413,7 +419,7 @@ def test_find_interesting_entities_by_record_id_bad_record_id(
         _ = sz_engine.find_interesting_entities_by_record_id(data_source_code, bad_record_id, flags)
 
 
-def test_find_network_by_entity_id(sz_engine: SzEngineTest) -> None:
+def test_find_network_by_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_network_by_entity_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -437,7 +443,7 @@ def test_find_network_by_entity_id(sz_engine: SzEngineTest) -> None:
     assert schema(network_schema) == actual_as_dict
 
 
-def test_find_network_by_entity_id_bad_entity_ids(sz_engine: SzEngineTest) -> None:
+def test_find_network_by_entity_id_bad_entity_ids(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_network_by_entity_id()."""
     bad_entity_list = [0, 1]
     max_degrees = 2
@@ -448,7 +454,7 @@ def test_find_network_by_entity_id_bad_entity_ids(sz_engine: SzEngineTest) -> No
         _ = sz_engine.find_network_by_entity_id(bad_entity_list, max_degrees, build_out_degrees, max_entities, flags)
 
 
-def test_find_network_by_record_id(sz_engine: SzEngineTest) -> None:
+def test_find_network_by_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_network_by_record_id()."""
     record_list: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -466,7 +472,7 @@ def test_find_network_by_record_id(sz_engine: SzEngineTest) -> None:
 
 
 def test_find_network_by_record_id_bad_data_source_code(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().find_network_by_record_id()."""
     bad_record_list: List[Tuple[str, str]] = [
@@ -481,7 +487,7 @@ def test_find_network_by_record_id_bad_data_source_code(
         _ = sz_engine.find_network_by_record_id(bad_record_list, max_degrees, build_out_degrees, max_entities, flags)
 
 
-def test_find_network_by_record_id_bad_record_ids(sz_engine: SzEngineTest) -> None:
+def test_find_network_by_record_id_bad_record_ids(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_network_by_record_id()."""
     bad_record_list: List[Tuple[str, str]] = [
         ("CUSTOMERS", "9999"),
@@ -495,7 +501,7 @@ def test_find_network_by_record_id_bad_record_ids(sz_engine: SzEngineTest) -> No
         _ = sz_engine.find_network_by_record_id(bad_record_list, max_degrees, build_out_degrees, max_entities, flags)
 
 
-def test_find_path_by_entity_id(sz_engine: SzEngineTest) -> None:
+def test_find_path_by_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_path_by_entity_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -521,7 +527,7 @@ def test_find_path_by_entity_id(sz_engine: SzEngineTest) -> None:
     assert schema(path_schema) == actual_as_dict
 
 
-def test_find_path_by_entity_id_bad_entity_ids(sz_engine: SzEngineTest) -> None:
+def test_find_path_by_entity_id_bad_entity_ids(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_path_by_entity_id()."""
     bad_start_entity_id = 0
     bad_end_entity_id = 1
@@ -541,7 +547,7 @@ def test_find_path_by_entity_id_bad_entity_ids(sz_engine: SzEngineTest) -> None:
         )
 
 
-def test_find_path_by_record_id(sz_engine: SzEngineTest) -> None:
+def test_find_path_by_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_path_by_record_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -571,7 +577,7 @@ def test_find_path_by_record_id(sz_engine: SzEngineTest) -> None:
     assert schema(path_schema) == actual_as_dict
 
 
-def test_find_path_by_record_id_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_find_path_by_record_id_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_path_by_record_id()."""
     bad_start_data_source_code = "XXXX"
     start_record_id = "9999"
@@ -594,7 +600,7 @@ def test_find_path_by_record_id_bad_data_source_code(sz_engine: SzEngineTest) ->
         )
 
 
-def test_find_path_by_record_id_bad_record_ids(sz_engine: SzEngineTest) -> None:
+def test_find_path_by_record_id_bad_record_ids(sz_engine: SzEngine) -> None:
     """Test SzEngine().find_path_by_record_id()."""
     start_data_source_code = "CUSTOMERS"
     bad_start_record_id = "9999"
@@ -617,13 +623,13 @@ def test_find_path_by_record_id_bad_record_ids(sz_engine: SzEngineTest) -> None:
         )
 
 
-def test_get_active_config_id(sz_engine: SzEngineTest) -> None:
+def test_get_active_config_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_active_config_id()."""
     actual = sz_engine.get_active_config_id()
     assert actual >= 0
 
 
-def test_get_entity_by_entity_id(sz_engine: SzEngineTest) -> None:
+def test_get_entity_by_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_by_entity_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -638,7 +644,7 @@ def test_get_entity_by_entity_id(sz_engine: SzEngineTest) -> None:
     assert schema(resolved_entity_schema) == actual_as_dict
 
 
-def test_get_entity_by_record_id(sz_engine: SzEngineTest) -> None:
+def test_get_entity_by_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_by_record_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -653,7 +659,7 @@ def test_get_entity_by_record_id(sz_engine: SzEngineTest) -> None:
     assert schema(resolved_entity_schema) == actual_as_dict
 
 
-def test_get_entity_by_record_id_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_get_entity_by_record_id_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_by_record_id()."""
     bad_data_source_code = "XXXX"
     record_id = "9999"
@@ -662,7 +668,7 @@ def test_get_entity_by_record_id_bad_data_source_code(sz_engine: SzEngineTest) -
         _ = sz_engine.get_entity_by_record_id(bad_data_source_code, record_id, flags)
 
 
-def test_get_entity_by_record_id_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_get_entity_by_record_id_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_by_record_id()."""
     data_source_code = "CUSTOMERS"
     bad_record_id = "9999"
@@ -671,7 +677,7 @@ def test_get_entity_by_record_id_bad_record_id(sz_engine: SzEngineTest) -> None:
         _ = sz_engine.get_entity_by_record_id(data_source_code, bad_record_id, flags)
 
 
-def test_get_record(sz_engine: SzEngineTest) -> None:
+def test_get_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_record()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -686,7 +692,7 @@ def test_get_record(sz_engine: SzEngineTest) -> None:
     assert schema(record_schema) == actual_as_dict
 
 
-def test_get_record_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_get_record_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_record()."""
     bad_data_source_code = "XXXX"
     record_id = "9999"
@@ -695,7 +701,7 @@ def test_get_record_bad_data_source_code(sz_engine: SzEngineTest) -> None:
         _ = sz_engine.get_record(bad_data_source_code, record_id, flags)
 
 
-def test_get_record_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_get_record_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_record()."""
     data_source_code = "CUSTOMERS"
     bad_record_id = "9999"
@@ -704,7 +710,7 @@ def test_get_record_bad_record_id(sz_engine: SzEngineTest) -> None:
         _ = sz_engine.get_record(data_source_code, bad_record_id, flags)
 
 
-def test_get_redo_record(sz_engine: SzEngineTest) -> None:
+def test_get_redo_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_redo_record()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -718,14 +724,14 @@ def test_get_redo_record(sz_engine: SzEngineTest) -> None:
     assert schema(redo_record_schema) == actual_as_dict
 
 
-def test_get_stats(sz_engine: SzEngineTest) -> None:
+def test_get_stats(sz_engine: SzEngine) -> None:
     """Test SzEngine().stats()."""
     actual = sz_engine.get_stats()
     actual_as_dict = json.loads(actual)
     assert schema(stats_schema) == actual_as_dict
 
 
-def test_get_virtual_entity_by_record_id(sz_engine: SzEngineTest) -> None:
+def test_get_virtual_entity_by_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_virtual_entity_by_record_id()."""
     record_list: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -740,7 +746,7 @@ def test_get_virtual_entity_by_record_id(sz_engine: SzEngineTest) -> None:
 
 
 def test_get_virtual_entity_by_record_id_bad_data_source_code(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().get_virtual_entity_by_record_id()."""
     bad_record_list: List[Tuple[str, str]] = [
@@ -753,7 +759,7 @@ def test_get_virtual_entity_by_record_id_bad_data_source_code(
 
 
 def test_get_virtual_entity_by_record_id_bad_record_ids(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().get_virtual_entity_by_record_id()."""
     bad_record_list: List[Tuple[str, str]] = [
@@ -765,7 +771,7 @@ def test_get_virtual_entity_by_record_id_bad_record_ids(
         _ = sz_engine.get_virtual_entity_by_record_id(bad_record_list, flags)
 
 
-def test_how_entity_by_entity_id(sz_engine: SzEngineTest) -> None:
+def test_how_entity_by_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().how_entity_by_entity_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -779,7 +785,7 @@ def test_how_entity_by_entity_id(sz_engine: SzEngineTest) -> None:
     assert schema(how_results_schema) == actual_as_dict
 
 
-def test_how_entity_by_entity_id_bad_entity_id(sz_engine: SzEngineTest) -> None:
+def test_how_entity_by_entity_id_bad_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().how_entity_by_entity_id()."""
     bad_entity_id = 0
     flags = SzEngineFlags.SZ_HOW_ENTITY_DEFAULT_FLAGS
@@ -787,7 +793,7 @@ def test_how_entity_by_entity_id_bad_entity_id(sz_engine: SzEngineTest) -> None:
         _ = sz_engine.how_entity_by_entity_id(bad_entity_id, flags)
 
 
-def test_preprocess_record(sz_engine: SzEngineTest) -> None:
+def test_preprocess_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().preprocess_record()."""
     record_definition: str = DATA_SOURCES.get("CUSTOMERS", {}).get("1001", {}).get("Json", {})
     flags = SzEngineFlags.SZ_RECORD_DEFAULT_FLAGS
@@ -803,12 +809,12 @@ def test_preprocess_record(sz_engine: SzEngineTest) -> None:
 #         sz_engine.preprocess_record(json.dumps(record_definition))
 
 
-def test_prime_engine(sz_engine: SzEngineTest) -> None:
+def test_prime_engine(sz_engine: SzEngine) -> None:
     """Test SzEngine().prime_engine()."""
     sz_engine.prime_engine()
 
 
-def test_reevaluate_entity(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_entity(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_id_from_record_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -820,14 +826,14 @@ def test_reevaluate_entity(sz_engine: SzEngineTest) -> None:
     delete_records(sz_engine, test_records)
 
 
-def test_reevaluate_entity_bad_entity_id(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_entity_bad_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_id_from_record_id()."""
     bad_entity_id = 0
     flags = SZ_WITHOUT_INFO
     sz_engine.reevaluate_entity(bad_entity_id, flags)
 
 
-def test_reevaluate_entity_with_info(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_entity_with_info(sz_engine: SzEngine) -> None:
     """Test SzEngine().reevaluate_entity_with_info()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -841,14 +847,14 @@ def test_reevaluate_entity_with_info(sz_engine: SzEngineTest) -> None:
     assert schema(add_record_with_info_schema) == actual_as_dict
 
 
-def test_reevaluate_entity_with_info_bad_entity_id(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_entity_with_info_bad_entity_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().reevaluate_entity_with_info()."""
     bad_entity_id = 0
     flags = SzEngineFlags.SZ_WITH_INFO
     _ = sz_engine.reevaluate_entity(bad_entity_id, flags)
 
 
-def test_reevaluate_record(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().get_entity_id_from_record_id()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -861,7 +867,7 @@ def test_reevaluate_record(sz_engine: SzEngineTest) -> None:
     delete_records(sz_engine, test_records)
 
 
-def test_reevaluate_record_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_record_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().reevaluate_record()."""
     bad_data_source_code = "XXXX"
     record_id = "9999"
@@ -870,7 +876,7 @@ def test_reevaluate_record_bad_data_source_code(sz_engine: SzEngineTest) -> None
         sz_engine.reevaluate_record(bad_data_source_code, record_id, flags)
 
 
-def test_reevaluate_record_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_record_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().reevaluate_record()."""
     data_source_code = "CUSTOMERS"
     bad_record_id = "9999"
@@ -883,7 +889,7 @@ def test_reevaluate_record_bad_record_id(sz_engine: SzEngineTest) -> None:
 #     sz_engine.reevaluate_record(data_source_code, bad_record_id, flags)
 
 
-def test_reevaluate_record_with_info(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_record_with_info(sz_engine: SzEngine) -> None:
     """Test SzEngine().reevaluate_entity_with_info()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -899,7 +905,7 @@ def test_reevaluate_record_with_info(sz_engine: SzEngineTest) -> None:
 
 
 def test_reevaluate_record_with_info_bad_data_source_code(
-    sz_engine: SzEngineTest,
+    sz_engine: SzEngine,
 ) -> None:
     """Test SzEngine().reevaluate_entity_with_info()."""
     bad_data_source_code = "XXXX"
@@ -909,7 +915,7 @@ def test_reevaluate_record_with_info_bad_data_source_code(
         _ = sz_engine.reevaluate_record(bad_data_source_code, record_id, flags)
 
 
-def test_reevaluate_record_with_info_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_reevaluate_record_with_info_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().reevaluate_entity_with_info()."""
     data_source_code = "CUSTOMERS"
     bad_record_id = "9999"
@@ -920,7 +926,7 @@ def test_reevaluate_record_with_info_bad_record_id(sz_engine: SzEngineTest) -> N
     #     _ = sz_engine.reevaluate_record(data_source_code, bad_record_id, flags)
 
 
-def test_search_by_attributes(sz_engine: SzEngineTest) -> None:
+def test_search_by_attributes(sz_engine: SzEngine) -> None:
     """Test SzEngine().search_by_attributes
     ()."""
     test_records: List[Tuple[str, str]] = [
@@ -939,7 +945,7 @@ def test_search_by_attributes(sz_engine: SzEngineTest) -> None:
         assert schema(search_schema) == actual_as_dict
 
 
-def test_search_by_attributes_bad_attributes(sz_engine: SzEngineTest) -> None:
+def test_search_by_attributes_bad_attributes(sz_engine: SzEngine) -> None:
     """Test SzEngine().search_by_attributes()."""
     bad_attributes = "{"
     search_profile = ""
@@ -948,7 +954,7 @@ def test_search_by_attributes_bad_attributes(sz_engine: SzEngineTest) -> None:
         _ = sz_engine.search_by_attributes(bad_attributes, flags, search_profile)
 
 
-def test_why_entities(sz_engine: SzEngineTest) -> None:
+def test_why_entities(sz_engine: SzEngine) -> None:
     """Test SzEngine().why_entities()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -964,7 +970,7 @@ def test_why_entities(sz_engine: SzEngineTest) -> None:
     assert schema(why_entities_results_schema) == actual_as_dict
 
 
-def test_why_entities_bad_entity_ids(sz_engine: SzEngineTest) -> None:
+def test_why_entities_bad_entity_ids(sz_engine: SzEngine) -> None:
     """Test SzEngine().why_entities()."""
     bad_entity_id_1 = 0
     entity_id_2 = 1
@@ -978,7 +984,7 @@ def test_why_record_in_entity() -> None:
     # TODO: implement.
 
 
-def test_why_records(sz_engine: SzEngineTest) -> None:
+def test_why_records(sz_engine: SzEngine) -> None:
     """Test SzEngine().why_records()."""
     test_records: List[Tuple[str, str]] = [
         ("CUSTOMERS", "1001"),
@@ -996,7 +1002,7 @@ def test_why_records(sz_engine: SzEngineTest) -> None:
     assert schema(why_entity_results_schema) == actual_as_dict
 
 
-def test_why_records_bad_data_source_code(sz_engine: SzEngineTest) -> None:
+def test_why_records_bad_data_source_code(sz_engine: SzEngine) -> None:
     """Test SzEngine().why_records()."""
     data_source_code_1 = "CUSTOMERS"
     record_id_1 = "1001"
@@ -1007,7 +1013,7 @@ def test_why_records_bad_data_source_code(sz_engine: SzEngineTest) -> None:
         _ = sz_engine.why_records(data_source_code_1, record_id_1, bad_data_source_code_2, record_id_2, flags)
 
 
-def test_why_records_bad_record_id(sz_engine: SzEngineTest) -> None:
+def test_why_records_bad_record_id(sz_engine: SzEngine) -> None:
     """Test SzEngine().why_records()."""
     data_source_code_1 = "CUSTOMERS"
     record_id_1 = "1001"
@@ -1027,7 +1033,7 @@ def test_add_record_using_context_managment() -> None:
     """Test the use of SzEngine in context."""
     grpc_url = "localhost:8261"
     grpc_channel = grpc.insecure_channel(grpc_url)
-    with SzEngineTest(grpc_channel=grpc_channel) as sz_engine:
+    with SzEngine(grpc_channel=grpc_channel) as sz_engine:
         data_source_code = "TEST"
         record_id = "2"
         record_definition = "{}"
@@ -1039,11 +1045,11 @@ def test_constructor() -> None:
     """Test constructor."""
     grpc_url = "localhost:8261"
     grpc_channel = grpc.insecure_channel(grpc_url)
-    actual = SzEngineTest(grpc_channel=grpc_channel)
-    assert isinstance(actual, SzEngineTest)
+    actual = SzEngine(grpc_channel=grpc_channel)
+    assert isinstance(actual, SzEngine)
 
 
-def test_process_redo_record(sz_engine: SzEngineTest) -> None:
+def test_process_redo_record(sz_engine: SzEngine) -> None:
     """Test SzEngine().prime_engine()."""
     flags = SZ_WITHOUT_INFO
     while sz_engine.count_redo_records() > 0:
@@ -1115,7 +1121,7 @@ def test_process_redo_record(sz_engine: SzEngineTest) -> None:
 # -----------------------------------------------------------------------------
 
 
-def add_records(sz_engine: SzEngineTest, record_id_list: List[Tuple[str, str]]) -> None:
+def add_records(sz_engine: SzEngine, record_id_list: List[Tuple[str, str]]) -> None:
     """Add all of the records in the list."""
     flags = SZ_WITHOUT_INFO
     for record_identification in record_id_list:
@@ -1130,7 +1136,7 @@ def add_records(sz_engine: SzEngineTest, record_id_list: List[Tuple[str, str]]) 
         )
 
 
-def add_records_truthset(sz_engine: SzEngineTest, do_redo: bool = True) -> None:
+def add_records_truthset(sz_engine: SzEngine, do_redo: bool = True) -> None:
     """Add all truth-set the records."""
     flags = SZ_WITHOUT_INFO
     for record_set in DATA_SOURCES.values():
@@ -1147,7 +1153,7 @@ def add_records_truthset(sz_engine: SzEngineTest, do_redo: bool = True) -> None:
             sz_engine.process_redo_record(redo_record, flags)
 
 
-def delete_records(sz_engine: SzEngineTest, record_id_list: List[Tuple[str, str]]) -> None:
+def delete_records(sz_engine: SzEngine, record_id_list: List[Tuple[str, str]]) -> None:
     """Delete all of the records in the list."""
     flags = SZ_WITHOUT_INFO
     for record_identification in record_id_list:
@@ -1157,7 +1163,7 @@ def delete_records(sz_engine: SzEngineTest, record_id_list: List[Tuple[str, str]
         sz_engine.delete_record(record.get("DataSource", ""), record.get("Id", ""), flags)
 
 
-def delete_records_truthset(sz_engine: SzEngineTest) -> None:
+def delete_records_truthset(sz_engine: SzEngine) -> None:
     """Delete all truth-set the records."""
     flags = SZ_WITHOUT_INFO
     for record_set in DATA_SOURCES.values():
@@ -1165,7 +1171,7 @@ def delete_records_truthset(sz_engine: SzEngineTest) -> None:
             sz_engine.delete_record(record.get("DataSource"), record.get("Id"), flags)
 
 
-def get_entity_id_from_record_id(sz_engine: SzEngineTest, data_source_code: str, record_id: str) -> int:
+def get_entity_id_from_record_id(sz_engine: SzEngine, data_source_code: str, record_id: str) -> int:
     """Given a (datasource, record_id), return the entity ID."""
     entity = sz_engine.get_entity_by_record_id(data_source_code, record_id)
     entity_as_dict = json.loads(entity)
@@ -1185,30 +1191,30 @@ def szconfig_fixture() -> SzConfig:
     """
     grpc_url = "localhost:8261"
     grpc_channel = grpc.insecure_channel(grpc_url)
-    result = SzConfigTest(grpc_channel=grpc_channel)
+    result = SzConfig(grpc_channel=grpc_channel)
     return result
 
 
 @pytest.fixture(name="sz_configmanager", scope="function")
-def szconfigmanager_fixture() -> SzConfigManagerTest:
+def szconfigmanager_fixture() -> SzConfigManager:
     """
     Single szconfigmanager object to use for all tests.
     """
     grpc_url = "localhost:8261"
     grpc_channel = grpc.insecure_channel(grpc_url)
-    result = SzConfigManagerTest(grpc_channel=grpc_channel)
+    result = SzConfigManager(grpc_channel=grpc_channel)
     return result
 
 
 @pytest.fixture(name="sz_engine", scope="function")
-def szengine_fixture() -> SzEngineTest:
+def szengine_fixture() -> SzEngine:
     """
     Single SzEngine object to use for all tests.
     """
 
     grpc_url = "localhost:8261"
     grpc_channel = grpc.insecure_channel(grpc_url)
-    result = SzEngineTest(grpc_channel=grpc_channel)
+    result = SzEngine(grpc_channel=grpc_channel)
     return result